requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0
redis[hiredis]>=5.0.1
celery[redis]>=5.3.6
pyyaml>=6.0.1
litellm>=1.15.0
tiktoken>=0.5.2
minio>=7.2.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
backoff>=2.2.1
tenacity>=8.2.3
chromadb>=0.4.22
